Generate interactive studio for your model

The main goal of this function is to connect two local model explainers: Ceteris Paribus and Break Down. It also shows global explainers for your model such as Partial Dependency and Feature Importance.

modelStudio(x, ...)

# S3 method for explainer
modelStudio(x, new_observation, facet_dim = c(2, 2),
  max_features = 10, N = 500, ...)

# S3 method for default
modelStudio(x, new_observation, facet_dim = c(2, 2),
  max_features = 10, N = 500, data, y, predict_function = predict,
  label = NULL, ...)

Arguments

x
an explainer created with function `DALEX::explain()` or a model to be explained.

...
other parameters.

new_observation
a new observation with columns that correspond to variables used in the model.

facet_dim
dimensions of the grid. Default is 2x2.

max_features
maximal number of features to be included in BreakDown and FeatureImportance plot.

N
number of observations used for calculation of partial dependency profiles. Default is 500.

data
validation dataset, will be extracted from `x` if it's an explainer.

y
true labels for `data`, will be extracted from `x` if it's an explainer.

predict_function
predict function, will be extracted from `x` if it's an explainer.

label
a name of the model, will be extracted from `x` if it's an explainer.

Value

an object of the `r2d3` class.

References

ingredients https://modeloriented.github.io/ingredients/ iBreakDown https://modeloriented.github.io/iBreakDown/

Examples

library("dime")
library("DALEX")

#> Welcome to DALEX (version: 0.4.5).
#> Find examples and detailed introduction at: https://pbiecek.github.io/PM_VEE/
#> Additional features will be available after installation of: ALEPlot, breakDown, pdp, factorMerger.
#> Use 'install_dependencies()' to get all suggested dependencies


titanic <- na.omit(titanic)
set.seed(1313)
titanic_small <- titanic[sample(1:nrow(titanic), 500), c(1,2,3,6,7,9)]

model_titanic_glm <- glm(survived == "yes" ~ gender + age + fare + class + sibsp,
                         data = titanic_small, family = "binomial")

explain_titanic_glm <- explain(model_titanic_glm,
                               data = titanic_small[,-6],
                               y = titanic_small$survived == "yes",
                               label = "glm")

#> Preparation of a new explainer is initiated
#>   -> model label       :  glm 
#>   -> data              :  500  rows  5  cols 
#>   -> target variable   :  500  values 
#>   -> predict function  :  yhat.glm  will be used (default)
#>   -> predicted values  :  numerical, min =  0.001319881 , mean =  0.298 , max =  0.9741794  
#>   -> residual function :  difference between y and yhat (default)
#>   -> residuals         :  numerical, min =  -0.9083374 , mean =  -2.164108e-10 , max =  0.9628104  
#> A new explainer has been created!


new_observation <- titanic_small[1:10,-6]

modelStudio(explain_titanic_glm, new_observation[1,])

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

modelStudio(explain_titanic_glm, new_observation, N = 50)

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%