Generates interactive studio to explain predictive model

This tool uses your model, data and new observations, to provide local and global explanations. It generates plots and descriptions in the form of serverless HTML site, that supports animations and interactivity made with D3.js.

modelStudio(x, ...)

# S3 method for explainer
modelStudio(x, new_observation, facet_dim = c(2, 2),
  time = 500, max_features = 10, N = 500, B = 25,
  show_info = TRUE, parallel = FALSE, options = NULL, ...)

# S3 method for default
modelStudio(x, new_observation, facet_dim = c(2, 2),
  time = 500, max_features = 10, N = 500, B = 25,
  show_info = TRUE, parallel = FALSE, options = NULL, data, y,
  predict_function = predict, label = NULL, ...)

Arguments

x
an explainer created with function DALEX::explain() or a model to be explained.

...
other parameters.

new_observation
a new observation with columns that correspond to variables used in the model.

facet_dim
dimensions of the grid. Default is 2x2.

time
in ms. Set animation length. Default is 1000.

max_features
maximum number of features to be included in Break Down and Shapley Values plot.

N
number of observations used for calculation of partial dependency profiles. Default is 500.

B
number of random paths used for calculation of shapley values. Default is 25.

show_info
verbose progress bar on console? Default is TRUE.

parallel
speed up computation using parallelMap::parallelMap(). Default is FALSE.

options
customize modelStudio. See getOptions and vignette.

data
validation dataset, will be extracted from x if it's an explainer.

y
true labels for data, will be extracted from x if it's an explainer.

predict_function
predict function, will be extracted from x if it's an explainer.

label
a name of the model, will be extracted from x if it's an explainer.

Value

an object of the r2d3 class

References

ingredients iBreakDown DALEX DALEXtra

Examples

library("dime")

# ex1 classification

titanic <- na.omit(DALEX::titanic)
titanic_small <- titanic[,c(1,2,3,6,7,9)]

model_titanic_glm <- glm(survived == "yes" ~ gender + age + fare + class + sibsp,
                         data = titanic_small, family = "binomial")

explain_titanic_glm <- DALEX::explain(model_titanic_glm,
                                      data = titanic_small[,-6],
                                      y = titanic_small$survived == "yes",
                                      label = "glm",
                                      verbose = FALSE)

new_observations <- titanic_small[1:4,-6]
rownames(new_observations) <- c("Lucas","James", "Thomas", "Nancy")

modelStudio(explain_titanic_glm, new_observations,
            N = 100, B = 15, show_info = FALSE)

# ex2 regression

apartments <- DALEX::apartments

model_apartments <- glm(m2.price ~. ,
                        data = apartments)

explain_apartments <- DALEX::explain(model_apartments,
                                     data = apartments[,-1],
                                     y = apartments[,1],
                                     verbose = FALSE)

new_apartments <- apartments[1:2,-1]
rownames(new_apartments) <- c("ap1","ap2")

modelStudio(explain_apartments, new_apartments,
            facet_dim = c(1,2), time = 1000,
            N = 100, B = 15, show_info = FALSE)